refactor:code clean up
