+++ force-app/main/default/lwc/questionCard/questionCard.html
@@ -34,7 +34,7 @@
       <span class="required">*</span>
     </template>
 
-    <span>{question.Label__c}</span>
+    <span>{question.Position__c}. {question.Label__c}</span>
     <br />
 
     <lightning-badge label={question.Type__c}></lightning-badge>
