+++ force-app/main/default/lwc/questionForm/questionForm.html
@@ -1,5 +1,5 @@
 <template>
-  <div class="slds-box slds-box_x-small slds-m-around_x-small">
+  <div class="slds-box slds-box_x-small slds-m-top_x-small">
     <template if:true={isEditMode}>
       <h2>{label.edit_question_form_title}</h2>
     </template>
