--- force-app/main/default/lwc/questionPagination/labels.js
@@ -1,15 +1,0 @@
-import first_page from '@salesforce/label/c.first_page';
-import previous_page from '@salesforce/label/c.previous_page';
-import next_page from '@salesforce/label/c.next_page';
-import last_page from '@salesforce/label/c.last_page';
-
-const label = {
-  first_page,
-  previous_page,
-  next_page,
-  last_page
-}
-
-export {
-  label
-}
