+++ force-app/main/default/lwc/questionScreen/labels.js
@@ -1,34 +1,20 @@
-import no_questions from "@salesforce/label/c.no_questions";
 import previous from "@salesforce/label/c.previous";
 import next from "@salesforce/label/c.next";
 import no_template from "@salesforce/label/c.no_template"; 
 import unable_to_continue from "@salesforce/label/c.unable_to_continue";
 import limit_question_sexceeded from "@salesforce/label/c.limit_question_sexceeded"; 
 import error_message from "@salesforce/label/c.errorMessage";
-import you_can_create from "@salesforce/label/c.you_can_create";
-import questions from "@salesforce/label/c.questions";
-import question from "@salesforce/label/c.question";
-import can_no_longer_create_questions from "@salesforce/label/c.can_no_longer_create_questions";
 import you_must_have_at_least from "@salesforce/label/c.you_must_have_at_least";
-import find from "@salesforce/label/c.find";
-import clear from "@salesforce/label/c.clear";
-
 
 const label = {
-  no_questions,
   error_message,
   previous,
   next, 
   no_template, 
   unable_to_continue,
   limit_question_sexceeded,
-  you_can_create,
-  questions,
-  question,
-  can_no_longer_create_questions,
   you_must_have_at_least,
-  find,
-  clear
+  
 };
 
 export {
